Moved KPIComponent to plugins.
