+++ tools/gulptasks/dashboards/api-docs.json
@@ -12,7 +12,6 @@
         "../../../ts/Dashboards/Components/ComponentOptions.ts",
         "../../../ts/Dashboards/Components/EditableOptions.ts",
         "../../../ts/Dashboards/Components/HTMLComponent.ts",
-        "../../../ts/Dashboards/Components/KPIComponent.ts",
         "../../../ts/Dashboards/Components/Sync/Sync.ts",
         "../../../ts/Dashboards/Components/TextOptions.d.ts",
         "../../../ts/Dashboards/EditMode/ConfirmationPopup.ts",
@@ -29,6 +28,7 @@
         "../../../ts/Dashboards/PluginHandler.ts",
         "../../../ts/Dashboards/Plugins/DataGridComponent.ts",
         "../../../ts/Dashboards/Plugins/HighchartsComponent.ts",
+        "../../../ts/Dashboards/Plugins/KPIComponent.ts",
         "../../../ts/Dashboards/Plugins/NavigatorComponentOptions.ts",
         "../../../ts/Data/Connectors/CSVConnectorOptions.d.ts",
         "../../../ts/Data/Connectors/JSONConnectorOptions.d.ts",
