+++ ts/Dashboards/Plugins/HighchartsPlugin.ts
@@ -24,7 +24,7 @@ import type { Highcharts as H } from './HighchartsTypes';
 
 import HighchartsComponent from './HighchartsComponent.js';
 import HighchartsSyncHandlers from './HighchartsSyncHandlers.js';
-import KPIComponent from '../Components/KPIComponent.js';
+import KPIComponent from './KPIComponent.js';
 import NavigatorComponent from './NavigatorComponent.js';
 
 /* *
