+++ ts/Dashboards/Plugins/KPISyncHandlers.ts
@@ -21,7 +21,7 @@
 
 import type DataCursor from '../../Data/DataCursor';
 import type Sync from '../Components/Sync/Sync';
-import type KPIComponent from '../Components/KPIComponent';
+import type KPIComponent from './KPIComponent';
 
 import U from '../../Core/Utilities.js';
 const { defined } = U;
